pyproject.toml
[project]
name = "blog-cover-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.51.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "gunicorn>=21.2.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-login>=0.6.3",
    "flask-bcrypt>=1.0.1",
    "psycopg2-binary>=2.9.9",
    "google-cloud-storage>=2.14.0",
]
